added a way to protect routes and server actions

+++ src/actions/actions.ts
@@ -1,9 +1,17 @@
 "use server";
 
 import { prisma } from "@/lib/db";
+import { getKindeServerSession } from "@kinde-oss/kinde-auth-nextjs/server";
 import { revalidatePath } from "next/cache";
+import { redirect } from "next/navigation";
 
 export async function addExpense(formData: FormData) {
+  const { isAuthenticated } = getKindeServerSession();
+
+  if (!(await isAuthenticated())) {
+    return redirect("/");
+  }
+
   await prisma.expense.create({
     data: {
       amount: Number(formData.get("amount")),
@@ -14,6 +22,11 @@ export async function addExpense(formData: FormData) {
 }
 
 export async function deleteExpense(id: number) {
+  const { isAuthenticated } = getKindeServerSession();
+
+  if (!(await isAuthenticated())) {
+    return redirect("/");
+  }
   await prisma.expense.delete({
     where: {
       id: id,
